fixed some build warnings in 64bit

+++ arch/x86_64/include/types-x86_64.h
@@ -38,7 +38,9 @@ typedef unsigned short uint16_t;
 typedef          short sint16_t;
 typedef unsigned char  uint8_t;
 typedef          char  sint8_t;
-typedef unsigned int    intptr_t;
+
+typedef unsigned long  long intptr_t;
+
 typedef s32 off_t;
 typedef u32 size_t;
 typedef s32 uid_t;
